Moved last CFamily rules-related class from Integration.csproj
* "Integration" still has CFamily binding-related code
