+++ src/Integration/Resources/Strings.resx
@@ -750,14 +750,6 @@ This is the general format to use when writing errors to output window or when t
     <value>Project-level binding is not required for project '{0}'</value>
     <comment>Output window message about projects that don't require binding</comment>
   </data>
-  <data name="CFamily_UnableToLoadConnectedModeSettings" xml:space="preserve">
-    <value>Unable to load/locate connected mode settings. Falling back on standalone mode settings.</value>
-    <comment>Output window message if in connected mode but couldn't find or load the connected mode settings file</comment>
-  </data>
-  <data name="CFamily_UsingConnectedModeSettings" xml:space="preserve">
-    <value>Using connected mode settings. User-specified settings in settings.json will be ignored.</value>
-    <comment>Output window message</comment>
-  </data>
   <data name="FailedToCheckIfTestProject" xml:space="preserve">
     <value>Error identifying project '{0}' as test project: {1}</value>
     <comment>Error message indicating that an error occured while checking if a project is a test project</comment>
